LinkUIrlDialog Moved to seperate file
Still needs more refactoring
